refactor(router): move the prefetch cache into client-utils
Also use optional chaining for the 404 redirect check, stop exporting
the manager type, and list the manager in the hmr effect deps.
