+++ packages/waku/src/router/client-utils/navigate.ts
@@ -163,8 +163,7 @@ export const deriveNav = (outcome: {
     destination.elements &&
     outcome.getServerRedirect(destination.elements, destination.route);
   const route = redirect || destination.route;
-  const mode =
-    redirect && redirect.path === '/404' ? undefined : outcome.history;
+  const mode = redirect?.path === '/404' ? undefined : outcome.history;
   const url = redirect
     ? undefined
     : followed
