+++ packages/waku/tests/router-client.test.tsx
@@ -49,7 +49,7 @@ import {
   ROUTE_ID,
   decodeRoutePath,
 } from '../src/router/isomorphic-utils/route-path.js';
-import { PREFETCH_LIMIT } from '../src/router/prefetch-cache.js';
+import { PREFETCH_LIMIT } from '../src/router/client-utils/prefetch-cache.js';
 
 const postsSearchCodec = {
   id: 'posts-test',
